requirements.txt
# Django modules
django
django-cors-headers
djangorestframework
djangorestframework-jwt
djangorestframework-msgpack
drf-nested-routers
djangorestframework-filters==1.0.0.dev0
django-simple-history

# Generic Python modules
numpy
user_agents

# Formatting
yapf

# Image modules
opencv-python
opencv-contrib-python
Pillow

# Metadata modules
ExifRead
piexif
mutagen

# Heroku modules
# NOTE these are not actually required on non-heroku environments
gunicorn
django-heroku
